Refactor base page and do back button on blogpages
Split the base page show functions to their own file and add a back link
to the blogpages
