+++ assets/scripts/templates/blog-posts-with-buttons.handlebars
@@ -13,3 +13,5 @@
   </div>
   </div>
 {{/each}}
+
+<span class="page-links" id="back-link">Return to Main</span>
