+++ assets/scripts/templates/blog-posts.handlebars
@@ -7,3 +7,5 @@
       <p class="resource-body">{{blogPost.createdAt}}</p>
   </div>
 {{/each}}
+
+<span class="page-links" id="back-link">Return to Main</span>
